+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Fix pyflakes
+
 ### 0.10.2 - 2020-02-27
 
  - Fix serializers